+++ HISTORY.md
@@ -1,4 +1,12 @@
 
+v1.1
+---------------------
+
+  * removed delayAnima in favor of delay inside anima method
+  * now animations don't queue, they run asynchronously
+  * fixed bug on safary on stop() with jumptoend false
+  * removed options backfaceVisibility
+
 v1.0
 ---------------------